+++ src/Rask.Core/Components/H1.cs
@@ -1,6 +1,6 @@
 namespace Rask.Core.Components;
 
-public sealed class H1 : Element
+public sealed class H1 : HtmlHeadingElement
 {
     protected override string TagName => "h1";
 }