+++ src/Rask.Core/Components/H3.cs
@@ -1,6 +1,6 @@
 namespace Rask.Core.Components;
 
-public sealed class H3 : Element
+public sealed class H3 : HtmlHeadingElement
 {
     protected override string TagName => "h3";
 }