+++ src/Rask.Core/Components/Ins.cs
@@ -1,25 +1,6 @@
-using System.Text;
-
 namespace Rask.Core.Components;
 
-public sealed class Ins : Element
+public sealed class Ins : HtmlModElement
 {
     protected override string TagName => "ins";
-
-    public string? Cite { get; set; }
-    public string? DateTime { get; set; }
-
-    protected override void WriteAttributes(StringBuilder sb)
-    {
-        base.WriteAttributes(sb);
-        if (Cite is not null)
-        {
-            AppendUrlAttr(sb, "cite", Cite);
-        }
-
-        if (DateTime is not null)
-        {
-            AppendAttr(sb, "datetime", DateTime);
-        }
-    }
 }