+++ src/Rask.Core/Components/Tbody.cs
@@ -1,6 +1,6 @@
 namespace Rask.Core.Components;
 
-public sealed class Tbody : Element
+public sealed class Tbody : HtmlTableSectionElement
 {
     protected override string TagName => "tbody";
 }