+++ src/Rask.Core/Components/Tfoot.cs
@@ -1,6 +1,6 @@
 namespace Rask.Core.Components;
 
-public sealed class Tfoot : Element
+public sealed class Tfoot : HtmlTableSectionElement
 {
     protected override string TagName => "tfoot";
 }